Packet Tracer File
This is my .pkt file don;t worry everything has changed with my school regulatory so no one would have any problem and is approved by my Professor to share this on github. Also Don't copy all of it. I've put all the commands and the documentation in the repository and the project local file. Learn, improve and adapt my project if you have any problem accessing the router and the switches because we had to configure them all with encryption so be care-reading on the documentation. 

Thanks!

-Rexhino